update docs, move locker to haf.mark
