+++ docs/_templates/sidebarintro.html
@@ -8,6 +8,26 @@
   Haf is an easy http api test framework.
 </p>
 
+<p>
+  <a href="http://autotest.wang/">MainPage</a>
+</p>
+
+<p>
+  <a href="https://github.com/tsbxmw/haf/">GitHub</a>
+</p>
+
+<p>
+  <a href="https://pypi.org/project/haf/">Pypi haf</a>
+</p>
+
+<p>
+  <a href="https://github.com/tsbxmw/haf/">GitHub</a>
+</p>
+
+<p>
+  <a href="https://github.com/tsbxmw/haf/wiki/Quick-Start">Quick Start</a>
+</p>
+
 <div class="native-js">
   <div class="native-sponsor">Sponsored by #native_company# — Learn More</div>
   <a href="#native_link#" class="native-flex">
