+++ docs/_templates/sidebarlogo.html
@@ -1,63 +1,68 @@
 <p class="logo">
   <a href="{{ pathto(master_doc) }}">
-    <img class="logo" src="{{ pathto('_static/requests-sidebar.png', 1) }}" title="https://kennethreitz.org/tattoos"/>
+    <img class="logo" src="{{ pathto('_static/haf-sidebar.png', 1) }}" />
   </a>
 </p>
-<p>
-<iframe src="https://ghbtns.com/github-btn.html?user=requests&repo=requests&type=watch&count=true&size=large"
-  allowtransparency="true" frameborder="0" scrolling="0" width="200px" height="35px"></iframe>
-</p>
-
 
 <p>
-  Requests is an elegant and simple HTTP library for Python, built for
-  human beings. You are currently looking at the documentation of the
-  development release.
+  Haf is an easy http api test framework.
 </p>
 
-<p>Sponsored by <strong><a href="https://linode.com/">Linode</a></strong> and <a href="http://docs.python-requests.org/en/master/community/sponsors/#patron-sponsors">other wonderful organizations</a>.</p>
+<p>
+  <a href="http://autotest.wang/">MainPage</a>
+</p>
 
-<h3>Stay Informed</h3>
-<p>Receive updates on new releases and upcoming projects.</p>
+<p>
+  <a href="https://github.com/tsbxmw/haf/">GitHub</a>
+</p>
 
-<p><a href="https://tinyletter.com/kennethreitz">Join Mailing List</a>.</p>
+<p>
+  <a href="https://pypi.org/project/haf/">Pypi haf</a>
+</p>
 
-<hr/>
+<p>
+  <a href="https://github.com/tsbxmw/haf/">GitHub</a>
+</p>
 
-<script async type="text/javascript" src="//cdn.carbonads.com/carbon.js?zoneid=1673&serve=C6AILKT&placement=pythonrequestsorg" id="_carbonads_js"></script>
+<p>
+  <a href="https://github.com/tsbxmw/haf/wiki/Quick-Start">Quick Start</a>
+</p>
 
-<p><iframe src="https://ghbtns.com/github-btn.html?user=kennethreitz&type=follow&count=false"
-  allowtransparency="true" frameborder="0" scrolling="0" width="200" height="20"></iframe></p>
+<div class="native-js">
+  <div class="native-sponsor">Sponsored by #native_company# — Learn More</div>
+  <a href="#native_link#" class="native-flex">
+    <style>
+      .native-js {
+        background: linear-gradient(-30deg, #native_bg_color#E5, #native_bg_color#E5 45%, #native_bg_color# 45%) #fff;
+      }
 
-<p><a href="https://twitter.com/kennethreitz" class="twitter-follow-button" data-show-count="false">Follow @kennethreitz</a> <script>!function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0],p=/^http:/.test(d.location)?'http':'https';if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src=p+'://platform.twitter.com/widgets.js';fjs.parentNode.insertBefore(js,fjs);}}(document, 'script', 'twitter-wjs');</script></p>
+      .native-details,
+      .native-sponsor,
+      .native-bsa {
+        color: #native_color# !important;
+      }
 
+      .native-details:hover {
+        color: #native_color_hover# !important;
+      }
 
-<h3>Other Projects</h3>
+      .native-cta {
+        color: #native_cta_color#;
+        background-color: #native_cta_bg_color#;
+      }
 
-<p>More <a href="https://www.kennethreitz.org/">Kenneth Reitz</a> projects:</p>
-<ul>
-    <li><a href="https://html.python-requests.org/">Requests-HTML</a></li>
-    <li><a href="http://howtopython.org/">howtopython.org</a></li>
-    <li><a href="http://pipenv.org/">pipenv</a></li>
-    <li><a href="https://pep8.org/">pep8.org</a></li>
-    <li><a href="https://httpbin.org/">httpbin.org</a></li>
-    <li><a href="https://docs.python-guide.org/">The Python Guide</a></li>
-    <li><a href="https://github.com/kennethreitz/maya">Maya: Datetimes for Humans</a></li>
-    <li><a href="https://github.com/kennethreitz/records">Records: SQL for Humans</a></li>
-    <li><a href="http://www.git-legit.org">Legit: Git for Humans</a></li>
-    <li><a href="http://docs.python-tablib.org/en/latest/">Tablib: Tabular Datasets</a></li>
-</ul>
-
-
-<h3>Translations</h3>
-
-<ul>
-  <li><a href="http://docs.python-requests.org/">English</a></li>
-  <li><a href="http://fr.python-requests.org/">French</a></li>
-  <li><a href="http://de.python-requests.org/">German</a></li>
-  <li><a href="http://jp.python-requests.org/">Japanese</a></li>
-  <li><a href="http://cn.python-requests.org/">Chinese</a></li>
-  <li><a href="http://pt.python-requests.org/">Portuguese</a></li>
-  <li><a href="http://it.python-requests.org/">Italian</a></li>
-  <li><a href="http://es.python-requests.org/">Spanish</a></li>
-</ul>
+      .native-cta:hover {
+        color: #native_cta_color_hover;
+        background-color: #native_cta_bg_color_hover#;
+      }
+    </style>
+    <div class="native-main">
+      <img class="native-img" src="#native_logo#">
+      <div class="native-details">
+        <span class="native-company">#native_title#</span>
+        <span class="native-desc">#native_desc#</span>
+      </div>
+    </div>
+    <span class="native-cta">#native_cta#</span>
+  </a>
+</div>
