+++ haf/mark.py
@@ -1,7 +1,10 @@
 # encoding='utf-8'
 import functools
 import inspect
+import time
+
 from haf.case import PyCase
+from haf.common.lock import Lock
 
 
 class TestDecorator:
@@ -49,3 +52,44 @@ class ParameterizeDecorator:
 
 
 parameterize = ParameterizeDecorator
+
+
+class Locker:
+    def __init__(self, bus_client, key):
+        self.bus_client = bus_client
+        self.key = key
+
+    def get_lock(self):
+        if self.key == "result":
+            lock = self.bus_client.get_lock()
+        elif self.key == "web":
+            lock = self.bus_client.get_web_lock()
+        elif self.key == "case":
+            lock = self.bus_client.get_case_lock()
+        while True:
+            if not lock.empty():
+                return lock.get()
+            time.sleep(0.1)
+
+    def release_lock(self):
+        if self.key == "result":
+            return self.bus_client.get_lock().put(Lock)
+        elif self.key == "web":
+            return self.bus_client.get_web_lock().put(Lock)
+        elif self.key == "case":
+            return self.bus_client.get_case_lock().put(Lock)
+
+
+def locker(func):
+    @functools.wraps(func)
+    def lock(self, *args, **kwargs):
+        if len(args) > 0:
+            key = args[0]
+        else:
+            key = kwargs.get("key")
+        locker = Locker(self.bus_client, key)
+        locker.get_lock()
+        func(self, *args, **kwargs)
+        locker.release_lock()
+        return
+    return lock
