+++ haf/utils.py
@@ -19,57 +19,14 @@ from openpyxl import load_workbook
 from haf.config import *
 from haf.common.httprequest import HttpController
 from http.client import HTTPResponse
-from haf.common.lock import Lock
 from datetime import datetime
-from functools import wraps
 import random
 import platform
 import yaml
 
-from haf.mark import test, skip, parameterize, TestDecorator
+from haf.mark import test, skip
 
 logger = Log.getLogger(__name__)
-
-
-class Locker:
-    def __init__(self, bus_client, key):
-        self.bus_client = bus_client
-        self.key = key
-
-    def get_lock(self):
-        if self.key == "result":
-            lock = self.bus_client.get_lock()
-        elif self.key == "web":
-            lock = self.bus_client.get_web_lock()
-        elif self.key == "case":
-            lock = self.bus_client.get_case_lock()
-        while True:
-            if not lock.empty():
-                return lock.get()
-            time.sleep(0.1)
-
-    def release_lock(self):
-        if self.key == "result":
-            return self.bus_client.get_lock().put(Lock)
-        elif self.key == "web":
-            return self.bus_client.get_web_lock().put(Lock)
-        elif self.key == "case":
-            return self.bus_client.get_case_lock().put(Lock)
-
-
-def locker(func):
-    @wraps(func)
-    def lock(self, *args, **kwargs):
-        if len(args) > 0:
-            key = args[0]
-        else:
-            key = kwargs.get("key")
-        locker = Locker(self.bus_client, key)
-        locker.get_lock()
-        func(self, *args, **kwargs)
-        locker.release_lock()
-        return
-    return lock
 
 
 class Utils(object):
